!feat: use first class typescript enums  (#20)
* feat: use first class typescript enums

Enums are handled as first class types in Typescript. The previous behavior was seen as a more direct comparison of types, since Golang does not have native Enum support.

The previous behavior however was lossy, as it lost the enum member names. This new change is lossless, and the types can be reverted to the prior state.