chore: support attaching arbitrary comments to types (#57)
Adds a new type SupportsComments for handling comments on top level declarations and field properties.

The original idea was to create an AST node for comments. This works well when constructing an AST, but makes traversing an AST more difficult. The comment nodes add extra unbounded levels of indirection.

This solution keeps the comments on the related type.